[RM-5] Table Entity Creation

+++ src/main/java/com/jiinfotech/restomanager/db/dish/DishRepo.java
@@ -2,7 +2,7 @@ package com.jiinfotech.restomanager.db.dish;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DishRepo extends JpaRepository<Dish, Integer> {
+public interface DishRepo extends JpaRepository<Dish, Long> {
 
 }
